Created home page GUI
