+++ CS446Project/app/src/main/res/layout/music_player.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingBottom="5dp"
+    android:paddingLeft="5dp"
+    android:paddingRight="5dp"
+    android:paddingTop="5dp"
+    tools:context="com.example.qian.cs446project.MusicPlayer">
+
+    <RelativeLayout
+        android:id="@+id/buttonPanel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content">
+
+        <ImageView
+            android:id="@+id/playPauseButtons"
+            android:layout_marginTop="15dp"
+            android:layout_marginBottom="15dp"
+            android:src="@drawable/play"
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+            android:onClick="onTogglePlay" />
+
+        <ImageView
+            android:id="@+id/stopButton"
+            android:layout_marginTop="15dp"
+            android:layout_marginBottom="15dp"
+            android:layout_toEndOf="@id/playPauseButtons"
+            android:layout_toRightOf="@id/playPauseButtons"
+            android:src="@drawable/stop"
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+            android:onClick="onStop" />
+
+    </RelativeLayout>
+
+    <ListView
+        android:id="@+id/songlist"
+        android:layout_below="@id/buttonPanel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content">
+
+    </ListView>
+</RelativeLayout>
